Backend/Predictive-Maintenance/requirements.txt
appnope==0.1.2
autopep8==1.6.0
backcall==0.2.0
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
fonttools==4.28.2
ipykernel==6.6.0
ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.7.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
protobuf==3.19.1
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2